requirements.txt
# ── Core graph & ML ───────────────────────────────────────────────────────────
networkx==3.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
torch==2.3.1                    # required by sentence-transformers

# ── NLP ───────────────────────────────────────────────────────────────────────
spacy==3.7.5
# After install, also run: python -m spacy download en_core_web_sm

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib==3.9.1
pyvis==0.3.2

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy==1.26.4
tqdm==4.66.4
requests==2.32.3

# ── Notebook ──────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipywidgets==8.1.3

# ── Module 2 (install before running notebook 02) ────────────────────────────
python-louvain==0.16             # Louvain community detection
transformers==4.43.3             # HuggingFace (T5/BART summarization)

# ── Module 4 (install before running notebook 04) ────────────────────────────
anthropic>=0.28.0                # Claude Sonnet 4.6 — primary report generation backend
openai==1.37.1                   # secondary backend (optional)

# ── Web UI (Task 7) ───────────────────────────────────────────────────────────
fastapi==0.111.0                 # REST API backend
uvicorn[standard]==0.30.1        # ASGI server
python-multipart==0.0.9          # form data parsing (FastAPI dependency)
